linux: Remove warning about mixing enums and integers
On Fedora 38, with gcc (GCC) 13.2.1 20231011 (Red Hat 13.2.1-4), we get a
warning in conditional expressions mixing enums and integers:

```
../src/font.cc:181:46: warning: enumerated and non-enumerated type in conditional expression [-Wextra]
  181 |             size_t count = size < MAX_GLYPHS ? size : MAX_GLYPHS;
      |                            ~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~
```

+++ src/font.cc
@@ -178,7 +178,7 @@ struct font_cache
     {
         if (size)
         {
-            size_t count = size < MAX_GLYPHS ? size : MAX_GLYPHS;
+            size_t count = size < MAX_GLYPHS ? size : size_t(MAX_GLYPHS);
             data  *last  = cache + (size + ~0) % MAX_GLYPHS;
             data  *d     = last;
             for (size_t i = 0; i < count; i++)
